Whole bunch of stuff
Server now sends over all configuration to client on connection. We are now able to read each message individually! Connecting a client to a server will exchange messages rn and load server config into object in client for later use. Now we just need to follow the same format and send a json string for each packet, parse, and do whatev with it