Big features! Testing! Base 64 encoding! PNG files!
basically I got file transfer between server and client working on a low level, we just have to pound out the specifics of it. File name will be specified by server and that name will be sent to the client before initiating download. I tested with a text file and a png file. After the png file came through I compared the bytes to the original and it was completely in tact. I left a bunch of test stuff here, I will tidy it up before merging. Then we can start on building packets out of a loaded file, and the packet bodies will be build with b64. There might be some overhead with the packet headers since they are being converted into a string, but whatev ig